requirements.txt
psycopg2==2.6
celery==3.1.18
django-compressor==1.5
django-storages-redux==1.3
boto==2.38.0

isort==3.9.6
flake8==2.4.1

git+https://github.com/geodesign/django.git@geodesign_v2
git+https://github.com/geodesign/django-raster.git@master
git+https://github.com/geodesign/django-raster-aggregation.git@master
